Adapt ChimeraMap to generate multiple sequences.
Two parameters are added, one to specify the number of sequences to generate and one to specify the minimum number of different nt blocks required to encode a given aa block. The multiple sequences are then generated by concatenating the blocks, using a heuristic that alternates between two approaches in order to minimize long homologies between the sequences. Additionally, the max_pos filter is adapted and now tracks all the genes that contain any of the nt blocks (not just a single chosen one as previously).